refactor(encrypted-archive): move core code into `src/core` directory
