+++ packages/encrypted-archive/src/core/header/protocol-buffers-converter/simpleHeader.ts
