+++ packages/encrypted-archive/src/node/compress/gzip.ts
@@ -1,6 +1,6 @@
 import { createGunzip, createGzip } from 'zlib';
 
-import { GzipDisallowedOptions, gzipDisallowOptionNameList } from '../../types/compress';
+import { GzipDisallowedOptions, gzipDisallowOptionNameList } from '../../core/types/compress';
 import { validateDisallowedOptions } from './utils';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type, @typescript-eslint/explicit-module-boundary-types
