+++ packages/encrypted-archive/tests/helpers/stream.ts
@@ -1,8 +1,8 @@
 import * as stream from 'stream';
 import { promisify } from 'util';
 
-import type { StreamReaderInterface } from '../../src/utils/stream';
-import type { AsyncIterableIteratorReturn } from '../../src/utils/type';
+import type { AsyncIterableIteratorReturn } from '../../src/core/types/utils';
+import type { StreamReaderInterface } from '../../src/core/utils/stream';
 
 export const waitStreamFinished = promisify(stream.finished);
 export const pipelineAsync = promisify(stream.pipeline);
