+++ packages/encrypted-archive/tests/unit/key-derivation-function.ts
@@ -4,8 +4,8 @@ import * as util from 'util';
 import escapeStringRegexp from 'escape-string-regexp';
 
 import { getKDF } from '../../src/browser/key-derivation-function';
-import type { KeyDerivationOptions } from '../../src/types/key-derivation-function';
-import type { Argon2Options } from '../../src/types/key-derivation-function/argon2';
+import type { KeyDerivationOptions } from '../../src/core/types/key-derivation-function';
+import type { Argon2Options } from '../../src/core/types/key-derivation-function/argon2';
 import { addNegativeNumber, createDummySizeBuffer, rangeArray } from '../helpers';
 
 /** @see https://github.com/P-H-C/phc-winner-argon2/blob/16d3df698db2486dde480b09a732bf9bf48599f9/include/argon2.h#L57 */
